requirements.txt
Pillow==11.2.1
segno==1.6.6
